\name{myf-package}
\alias{myf-package}
\alias{myf}
\docType{package}
\title{
\packageTitle{myf}
}
\description{
\packageDescription{myf}
}
\details{

No description yet. 
Please be patient. 
The DESCRIPTION file:
\packageDESCRIPTION{myf}
\packageIndices{myf}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{myf}

Maintainer: \packageMaintainer{myf}
}

